Move RSpec helper methods into a separate support file
